+++ src/Simulation/dataManagement.h
@@ -27,7 +27,7 @@ public:
 
 	void setSampleDirectory(params& par, const std::string& );
 
-	static std::vector<value> getSample(const std::string& directory, float bounds);
+	static std::vector<value> getSample(const std::string& directory, double bounds);
 
 private:
