+++ src/Simulation/settings.h
@@ -20,7 +20,7 @@ struct params { //
     bool buildFromSample = false;
 
     int numSpecies = 300; // number of species  
-    float treeDensity = 0.04; // Number of trees per area or lambda for pois()
+    double treeDensity = 0.04; // Number of trees per area or lambda for pois()
 
     // Immigration settings
     bool metaCommunityImmigration = false;
@@ -29,23 +29,23 @@ struct params { //
     // Fragment interactions etc
     int numFragments = 3;
     bool equalFragmentSize = true;
-    float size = 100; // L ... Size of simulation Usefull for continuous 
-    std::vector<float> fragmentSizeList; // Is the equal Fragment size == false --> this this is used 
-    std::vector<std::vector<float>> nodeMap; // TODO: improve memeory useage << or decide if I need to because it's not used too much
+    double size = 100; // L ... Size of simulation Usefull for continuous 
+    std::vector<double> fragmentSizeList; // Is the equal Fragment size == false --> this this is used 
+    std::vector<std::vector<double>> nodeMap; // TODO: improve memeory useage << or decide if I need to because it's not used too much
 
     // Ecological settings
     bool fragmented = false; // Run with or without fragmentation
     bool neutralComp = true;
 
-    float searchArea = 20; // ZOI
-    float b1 = 6;
-    float b2 = 7;
-    float m = 0.001; // Immigration rate
-    float dispersalDis = 20;
-    float mort = 0.1; // mortality rate TODO : implement or remove 
-    float HNDD = 9; // Default HNDD strength TODO: implement
-    float CNDD = 9; // Default CNDD
-    float extinctionRate = 0.0001;
+    double searchArea = 20; // ZOI
+    double b1 = 6;
+    double b2 = 7;
+    double m = 0.001; // Immigration rate
+    double dispersalDis = 20;
+    double mort = 0.1; // mortality rate TODO : implement or remove 
+    double HNDD = 9; // Default HNDD strength TODO: implement
+    double CNDD = 9; // Default CNDD
+    double extinctionRate = 0.0001;
 
 
     std::string sampleDirectory = "";
@@ -85,8 +85,6 @@ private:
 
     void ResizeMatrix(int newSize);
 };
-
-
 
 #endif // !SPACE_SETTINGS
